+++ ktlint-ruleset-standard/src/test/kotlin/com/pinterest/ktlint/ruleset/standard/IndentationRuleTest.kt
@@ -3950,6 +3950,23 @@ internal class IndentationRuleTest {
         }
     }
 
+    @Test
+    fun `Issue 1644 - Given multiple nested brackets and some of them have ktlint-disable`() {
+        val code =
+            """
+            fun fooBar() {
+                fun foo() {
+                    // some code
+                } // ktlint-disable indent
+
+                fun bar() {
+                    // some code
+                }
+            }
+            """.trimIndent()
+        indentationRuleAssertThat(code).hasNoLintViolations()
+    }
+
     private companion object {
         val INDENT_STYLE_TAB = indentStyleProperty to PropertyType.IndentStyleValue.tab
     }